+++ documentation/docs/src/lib/components/Form/DatePicker/DatePicker.stories.svelte
@@ -44,6 +44,10 @@ import { CalendarDate, type DateValue } from '@internationalized/date';
       control: { type: 'text' },
       table: { category: 'API', subcategory: 'Base', defaultValue: { summary: 'dd/mm/yyyy' } },
     },
+    clearable: {
+      control: { type: 'boolean' },
+      table: { category: 'API', subcategory: 'Base' },
+    },
 
     // ------------------------------
     // State
@@ -96,6 +100,9 @@ import { CalendarDate, type DateValue } from '@internationalized/date';
       table: { category: 'API', subcategory: 'Slots' },
       description: 'Content rendered after the input',
     },
+    customTriggerIcon: {
+      table: { category: 'API', subcategory: 'Slots' },
+    },
   };
 
   // ------------------------------
@@ -116,6 +123,7 @@ import { CalendarDate, type DateValue } from '@internationalized/date';
   });
 
   let myValue = $state<DateValue>(new CalendarDate(2026, 4, 7));
+  let myValue2 = $state<DateValue>(new CalendarDate(2026, 12, 1));
 </script>
 
 <!-- ------------------------------ -->
@@ -128,6 +136,8 @@ import { CalendarDate, type DateValue } from '@internationalized/date';
   <DatePicker bind:value={myValue} />
 </Story>
 
+<Story name="Clearable" args={{ clearable: true, value: myValue2 }} />
+
 <Story name="Starts On Sunday" args={{ weekStartsOn: 0 }} />
 
 <Story name="Date Format" args={{ dateFormat: 'mm/dd/yyyy' }} />
@@ -135,6 +145,14 @@ import { CalendarDate, type DateValue } from '@internationalized/date';
 <Story name="Min Date" args={{ minValue: new CalendarDate(2026, 4, 7) }} />
 
 <Story name="Max Date" args={{ maxValue: new CalendarDate(2026, 4, 7) }} />
+
+<Story name="Custom Trigger Icon" asChild>
+  <DatePicker>
+    {#snippet customTriggerIcon()}
+      🗓️
+    {/snippet}
+  </DatePicker>
+</Story>
 
 <Story name="Light Theme" asChild>
   <Theme type="light">